Scripts/Configuration_Management/EA_Detailed_Design_Generator/Sources/requirements.txt
pyyaml
mdutils
tabulate
pywin32
xmltodict
pyinstaller